test: Merge connector integration tests
Merge the integration test of the outbound connector with the mock server into the other integration test.

Improve the test case by setting the URL as a connector secret.
